ttools: add code to prefer physical over logical plot fonts
CaptionerKeySet now has a setting that lets you prefer the
physical Lucida fonts (available in at least Oracle JREs) over the
logical font names (guaranteed available in all JREs).
The advantage of physical fonts is that they are JRE-independent.
At present this setting is switched off (so no change in previous
behaviour), but it may be set programmatically using a static member.