util: use AuthManager in URLDataSource and ContentCoding
UTIL now uses the AUTH package for transparent handling of
authentication.  URLDataSource uses the default AuthManager
when accessing remote URLs.

The handling of HTTP Basic Authentication in URLDataSource is
now withdrawn, since it would interfere with the AuthManager.
It might be possible to re-introduce similar functionality
into AuthManager at some point, but for now (the pretty dodgy
practice of) supplying user:password inline in the URL is
no longer possible.

Also as a convenience, ContentCoding now implements the
auth.UrlConnector interface, since it provides the relevant
functionality, and provides a new openStreamAuth convenience method.