vo: use AuthManager for authentication
All(?) appropriate access to URLs is now mediated using the
uk.ac.starlink.auth.AuthManager.  This should mean that authentication
is taken care of transparently.

This is a fairly fiddly business to get right, because as well as
(401) authentication negotiation, the HTTP handling required for
VO interactions includes various kinds of 3xx redirect handling.
By default, these are both managed by the HttpURLConnection,
but to insert our own authentication handling here we have to
bypass that and reconstruct the 401 and 3xx handling from scratch.
It is possible that new bugs or behaviour changes have been
inadvertently introduced; these will depend on the detailed
behaviour of VO services that this code interacts with, so may
be hard to spot.