topcat: use AuthManager for authentication
All(?) the appropriate places where URLs are interacted with are
now mediated by the uk.ac.starlink.auth.AuthManager, so that
VO-sensitive authentication ought to be taken care of transparently.

At this commit the TOPCAT application does not take advantage of this,
since no steps are taken to initialise the AuthManager.