ttools: STILTS uses AuthManager
The Stilts application now initialises the JVM-wide AuthManager
for console user interaction, so that authentication requests
should be intercepted and sent to the user where appropriate,
as far as they are implemented by the AUTH package.

Note that uk.ac.starlink.util.PropertyAuthenticator is no longer used,
so the star.basicauth.{user,password} properties can no longer be used
to set up BasicAuthentication from the command line.
This is replaced by the properties defined in
uk.ac.starlink.auth.UserInterface (auth.{username,password});
if they are defined, then those values will be used for all
authentication in STILTS.  This is somewhat dangerous (possibility
of leaking credentials to unintended services) so a WARNING is
issued through the logging system at application startup and
if properties-based authentication actually happens.

These changes are not so far reflected in the user documentation.