topcat: TOPCAT uses AuthManager
The Topcat application now initialises the JVM-wide AuthManager
for GUI interaction, so that authentication requests should be
intercepted and sent to the user where appropriate, as far as
they are implemented by the AUTH package.

The effect for HTTP Basic Authentication ought to be approximately
the same from a user point of view as before.

Note that the PropertyAuthenticator is no longer used, so it is
no longer possible to set Basic Authentication user/password
using system properties.

These changes are not so far reflected in the user documentation.