vo: Fix JSON example file
The JSON format is a bit picky when it comes to the use of colons (,):
they can be used only as a delimiter between items, but not after the
last item. While the normal json reader is sloppy here, the android
one (which we use in Debian due to license problems) is quite picky
and throws a JSONException here.

So, we just remove the colons at the end.