ecsv: avoid writing certain column metadata items
Do not write ColumnInfo.getAuxData items indiscriminately to
the per-column meta block of ECSV output.  Some items here are
useful, but others (e.g. the VOTable datatype attribute value)
are irrelevant or confusing.

I'm not sure whether it's best to blacklist or whitelist here,
but blacklist for now.  The approach and blacklist content may
need review with more examples.