table: defend against Healpix.getBestStartingDepth < 0
The cds-healpix-java function Healpix.getBestStartingDepth is
documented to return -1 for large input angles (>48).
I didn't realise that, and it was feeding invalid values to
other functions, causing (e.g.) sky matching to fail for match
radius greater than 6 degrees.

Modify the code to limit the value to a minimum of level 0
wherever getBestStartingDepth is invoked.