ttools: add JFreeSVG library v3.4
JFreeSVG is a library that implements Graphics2D with an SVG backend.
No dependencies, small size, nice.  It can be used for SVG output of
plots.

JFreeSVG 3.4 is the last version that targets a pre-modules JRE;
the next version is v4.1 which requires Java 11.  This library
is licenced under the GPL.

The file jfreesvg-3.4.jar was downloaded from
https://mvnrepository.com/artifact/org.jfree/jfreesvg/3.4.