ttools: avoid set null RenderingHints antialias values
There is a problem with save/restore of (at least) antialising
rendering hints in the JFreeSVG SVGGraphics2D implementation;
e.g. Graphics2D.getRenderingHint(RenderingHints.KEY_ANTIALIASING)
can return null, but attempting to restore that later by doing
Graphics2D.setRenderingHint(RenderingHints.KEY_ANTIALIASING,null)
throws an unchecked exception.  I don't know if that's really a bug,
but it's fairly easy to work round; preserve the state of the input
Graphics2D by cloning it and working on the clone rather than
saving and restoring hints on the input object.  None of the
places we need to change this is in a tight loop, so there shouldn't
be significant performance issues.

I thought about subclassing SVGGraphics2D and overriding
setRenderingHints with a fix instead, bub SVGGraphics2D is final,
so that's not possible.