table: StarTableFactory auto mode is now filename-sensitive
The makeStarTable methods in StarTableFactory that do not receive
a file format indicator will now fall back to looking at filenames
to guess file format (via the new TableBuilder.looksLikeFile method).

That means that you can now load a CSV file named xxx.csv without
having to explicitly say it's a CSV file.  People have been asking
for this for a long time, but I've resisted because I don't like
the idea of filenames determining file content, and I thought it
would be messy and inconsistent to do, as well as requiring changes
to the TableBuilder interface.  But it's not so bad.