ttools: rename SupplementSequence->SupplementData
This rather unimportant interface really isn't a sequence,
especially in view of recent changes elsewhere in the API,
so rename it.

I suspect that more tidying could be done by replacing it altogether
with use of some other existing interface or similar redesign of
the ColumnSupplement class, but don't attempt that now.