table: add StarTable.getRowAccess
Extend the StarTable interface with a getRowAccess method,
which provides a better way to do thread-safe table random access,
in the same way that getRowSequence does thread-safe sequential access.

This commit implements the method for all StarTables defined in the
code base, and represents a major backwardly-imcompatible change
to the StarTable interface.

Most uses of StarTables do not at this commit make use of this
new interface.