util: refactor SplittableSequence as Sequence
The SplittableSequence interface was just an extension of Splittable
that adds the next() method; remove it in favour of a new standalone
interface Sequence.  This is then added as a marker onto the interfaces
previously inheriting from SplittableSequence, as well as marking
RowSequence.  It can be used elsewhere later.

It's not really necessary to have this interface, the next method
could just be added to other interfaces anyway, but doing it like
this guarantees and documents a consistent contract across different
interfaces.