fits: add nom.tam.fits test for i10 format
I fixed a nom.tam.fits bug at 7f2d04319 which was giving int
rather than long values for ASCII table columns with i10 formatting.
This is a a (late) unit test for that bug fix.

The fix was in the starjava forked version of nom.tam.fits itself,
so if that tamfits is ever replaced by a more recent one,
this test will check that the behaviour is now correct.