topcat: fix plot subset name reuse regression bug
The commit f8ed61c29 (between topcat v4.7 and v4.7-1)
modified the way that subsets were identified when characterising
a plot, requiring a String rather than an Object.
That commit used the subset name as the string identifier,
meaning behaviour changed when a subset with a given name was
redefined - so if you created a subset by drawing a blob,
and then created a different blob subset reusing the same name,
the plot didn't realise that a replot was required and the
old subset content stayed in the plot display (until a replot
was forced).  In previous versions, a replot was triggered
automatically.

Fix it now by using a string that's unique to the subset instance
rather than the user-assigned subset name.  I don't think this
should break anything which wasn't broken in earlier versions,
since the equality behaviour will be the same as the RowSubset.equals
test that was there before; no (non-obsolete) RowSubset implementations
override Object.equals.

This issue was reported by Marina Kounkel on the topcat-user mailing list.