topcat: fix JComboBox generics regression bugs
When I updated most of the starjava classes to use generics in
Swing classes (around commit 56097187122, between topcat v4.7 and
v4.7-1), I changed some JComboBox-related code so that objects
of type JComboBox became JComboBox<T>, and when determining the
selected item, I changed in some places

   (T) comboBox.getSelectedItem()

to

   comboBox.getItemAt(comboBox.getSelectedIndex())

since it avoids the typecast.  However, these don't do exactly the
same thing!  In the case of an editable JComboBox (setEditable(true)),
the former form can return an object that has been added by the user,
but the latter form never will.  In practice this is most likely
to apply to String-typed combo boxes.

So, a few things were broken for editable combo boxes:

   - user-entered Vizier table names in the CDS XMatch Upload window
     (table could not be used)

   - user-entered UCD selection in synthetic column window
     (UCD was ignored)

   - user-entered row limit selection in Vizier load dialogue
     (table load failed)

Change the code back to something like its earlier state,
and fix these items.