topcat: ensure HiPS selection submenus are not too big
While I wasn't looking the number of Infrared HiPS services
increased so much that the popup menu for selecting IR HiPS
datasets got too tall to fit on the screen (of my laptop anyway),
which made it impossible to select items at the top and/or bottom
of the menu.

Add some code that breaks up sub-menus over a given size into
sub-sub-menus, so that all the items will fit on the screen
and can be selected.