table: move getAuxData from ColumnInfo to ValueInfo
The ValueInfo interface now has a getAuxData method, so that
auxiliary metadata can be stored wherever a ValueInfo is used
(e.g. table parameters) and not just in ColumnInfos.

The restriction of aux metadata to columns was a mistake in
the original API: there shouldn't be any difference between
columns and parameters in terms of what metadata can in principle
be stored.