table: move some AuxData access methods ColumnInfo->ValueInfo
The utility methods getAuxDatumByName and setAuxDatum have
moved up from ColumnInfo to ValueInfo, so that these can be
performed on any ValueInfo, not just a ColumnInfo.
These methods mirror the parameter access methods available
on StarTable.

Since these utility methods interact in predictable ways with the
getAuxData() method, implement them as default methods on the
interface rather than leaving them to subclasses to implement,
to underline this fact.
The actual implementation is done by the same utility methods
in the Tables class which do the same work for the corresponding
StarTable parameter access methods.