table: make StarTable parameter access methods default
The methods getParameterByName and setParameter on the StarTable
interface interact in predictable ways with the getParameters()
method, so implement these as default methods on the interface
rather than leaving them to subclasses to implement, to underline
this fact.  The actual implementation details (which don't do
anything complicated) are in the Tables utility class.