votable: extend aux metadata handling to PARAM not just FIELD
The various metadata items that can be specified on a VOTable FIELD
and turned into aux metadata items in the STIL API can now be
encoded for PARAMs as well, since ValueInfo now has getAuxData
in the same way as ColumnInfo always has.  In particular this means
that the TIMESYS and COOSYS metadata corresponding to PARAMs can
be recorded (and roundtripped) by STIL; previously this only
worked for FIELDs.