table: clone RowSequence.getRow() results in ListRowStore
The ListRowStore was storing Object[] results from RowSequence.getRow().
RowSequence.getRow() does not document whether that's OK or not,
i.e. whether the returned Object[] array is safe for reuse or whether
it may get scribbled over by subsequent getRow invocations.
Up till now, all getRow invocations have created a new row Object[]
array every time, but that is wasteful and it would be nice not to.
Since the behaviour is not documented (and ListRowStore is not
very heavily used for large datasets), get ListRowStore to be
more defensive.

I don't know whether there are other places where the assumption
is made that the Object[] arrays are safe for future use.
If not, maybe I should modify some of the RowSequence implementations
in the library code to reuse these arrays.