fits: lazily create ColFitsStarTable per-column Inputs
In ColFitsStarTable, create the per-column input objects only
when they are required, rather than pre-emptively for all columns,
in the RowSequence and RowAccess implementations.
Although this introduces an extra null test into each cell read
operation, this does not seem to impact the performance,
and it means that columns never needed are never (e.g.) mapped.