topcat: I/O handler docs now auto-generated in SUN/253
The user documentation for supported table input/output formats
is now generated automatically from the self-documenting
TableBuilder and StarTableWriter classes themselves.

This makes the documentation tidier and more maintainable;
previously this information was replicated
(more or less) in the STIL user document SUN/252, and
(in an abbreviated form) in the STILTS user document SUN/256.