fits: FITS TableWriters are now self-documenting
FitsPlusTableWriter and ColFitsPlusTableWriter now implement
DocumentedIOHandler.  The documentation itself is in the FITS package,
because that's where most of the functionality lives, but the
handlers supplying it are the *FitsPlusTableWriter implementations,
since those are the ones picked up by default by the output manager
(i.e. the fits-plus handler appears earlier in the handler list
than fits-basic, so if you write to x.fits with automatic format
guessing, you'll get a fits-plus file not a fits-basic one).