gbin: work around GaiaTools to improve exception logging
Logging exceptions is supposed to report them through the logging
system to the user, and if the -debug flag is supplied to topcat or
stilts, to provide a full stack trace.  That doesn't seem to work
in the GBIN package, I think because the GaiaTools classes it loads
are messing with the java.util.logging system configuration.

So work round it by using a custom logError method.
This should at least give users the thrown exception message and
class, which is probably enough for GBIN-aware users to work out
what's up.

If the GaiaTools classes are loaded, of course it means that
exception logging elsewhere in the STIL application will be
scuppered as well.  Too bad.