topcat: update MacOS packaging
Replace the build machinery used to prepare the TOPCAT .dmg files
for installation on Macs.

Before this I was using osxjarbundler, based on
https://github.com/tofi86/universalJavaApplicationStub.
That handled the messy business of finding either a java6 or a
java7+ installation on the MacOS machine at runtime;
these are installed in different places and different ways
since they are from different vendors (up to 1.6 was Apple,
1.7+ is Oracle).

Since the target platform for topcat is now java 8,
I don't need those complications (in fact they are a Bad Thing,
since finding a java6 can cause topcat to fail to run),
so revert to using appbundler, from
https://github.com/TheInfiniteKind/appbundler.
This is apparently an updated version of the older,
now abandoned? jarbundler.

Take the opportunity to make a few other improvements to the
MacOS install: include SUN/256 alongside SUN/253 and tidy up
some ancient stuff that was (I think) packaging demo files in an
unnecessary way.

Also move the bin and docs directories out of the top level and
put them only in TOPCAT.app/Contents/Resources, which seems to be
where MacOS expects that sort of thing to be.

I haven't tested this myself since I don't have a tame Mac,
but various users report it's basically working, though the
user needs to confirm a security exception when first loading
the .dmg file since I'm not an Apple developer.