table: add StarTable.getRowSplittable method
Add a method that lets tables control how splittable row sequences
are dispensed.  This is a good idea in that it allows upstream
insertion of progress monitoring, and in principle allows table
implementations with constraints on how tables should be split up
(e.g. not allowing a thousand accessors to run at different disk
positions simultaneously) to enforce them.

Implementing this method is not particularly easy for StarTables
in general; you can't just dispense an anonymous extension of a
WrapperRowSplittable class (no such class exists for this reason)
because such implementations generally have to have their own
named class so that they can dispense instances of that rather
than instances of the superclass.  That makes wrapper-type
StarTable implementation significantly harder than it was before.
You can just use the default AbstractStarTable implementation
(which calls Tables.getDefaultRowSplittable), but that doesn't
respect the splittable policy of upstream (wrapped) tables.
For the case where the rows are the same as the wrapped tables
and only the column structure changes, a utility class
MappingRowSplittable is provided.

This commit therefore provides non-trivial implementations of
getRowSplittable for many of the StarTable implementations,
but maybe not all.

Note that RowSplittables are supposed to provide the row index
of the current row.  This is necessary for some, though not most,
RowSplittable consumers, e.g. JEL evaluation including $index or
similar tokens, but it's problematic to provide in some implementations,
e.g. those with unpredictable row selection/inclusion semantics
and actual splitting.  The RowSplittable accommodates implementation
refusal to provide this information (rowIndex() may return null).
But some care is required on the part of the RowSplittable
consumers to ensure that where rowIndex is required, they make
sure to acquire a RowSplittable that is capable of yielding it
(which may mean eschewing a multithreaded implementation).
Some of the implementation in this commit is addressing that issue.

As part of this I improved the progress logging by using an external
timer.  It speeds progress-logged execution significantly.