ttools: add DocUtils.getXmlDescription(Documented) method
This extracts the XML documentation from an item that implements
the Documented interface, transforming it from XHTML-like form
to a form that can be inserted directly into the STILTS user document
(SUN).  A modest amount of XSLT transformation is performed.

The minimal fromXhtml method is used, which is by no means complete,
but it's sufficient for its required use of building the document
with the existing classes.