ttools: include example plotserver HTML and ipynb files
The error/documentation page returned by the PlotServlet now links
to working html and ipynb files that display use of the service
for basic but non-trivial plots.

An additional file ex-plots.html is also written at build time and
made available from the service; this contains all the Plot2Examples
as interactive plots in an HTML page.  This is not linked from
the service, since the required data will in general not be
available, but it's useful for testing; the server has to be
started up in the directory containing the test data, or otherwise
have it made available.