table: revert STIL version to 3.4-2
Despite the fact that the STIL version was previously bumped
to v3.4-2+, it really hasn't changed since the last release,
except for changes in table.join, which isn't really part of
the documented STIL library (and will be moved out of it
in future versions).  And a very tiny GBIN library change.
So leave the version the same as for the previous topcat/stilts release.