topcat: update JEL to v2.1.2
This JEL update has two changes:

  1. Fixes a bug: under some rather specific circumstances,
     attempting to compile a JEL expression with a forced expression
     type was resulting in a java.lang.VerifyError.

  2. Uses Boolean.valueOf() rather than new Boolean() to construct
     boolean instances, and similar behaviour for other wrapper types.
     That should result at least in some cases in better performance,
     or at least better memory management (fewer objects created).

The bug (1) manifested itself in a failure by TopcatCodec2
to re-ingest a session file with certain types of synthetic subset;
it was possible to work round that failure by invoking the
SyntheticColumn constructor in TopcatCodec2 with a null value
for class (rather than that obtained from the ColumnInfo),
but this JEL bugfix provides a better fix, provided in short
order by Konstantin Metlov.

As well as the JEL update, this commit includes a new test in
JELTest that would fail under the old JEL library.
I also added some missing test code (synthetic columns) in CodeTest
that related to this issue.

I compiled the jel_g.jar from source as usual (ant dist) to get a
version with a suitable classfile version.

jel_g.jar is also copied into the jsky JEL version.
This shouldn't really be necessary, and may not be desirable
(though as far as I know JEL backward compatibility is OK),
but if I don't do it then the topcat unit tests pick up the
wrong JEL version.  That ought to be fixable instead by adjusting
the test classpath in the topcat build.xml.  The correct solution
would be to put JEL in its own third-party package.