ttools: taplint copes with min detail from VOSI1.1 /tables
The /tables endpoint at VOSI 1.1 is permitted to return table
metadata lacking detail on columns and foreign keys (it can
just contain table-level metadata, generally if the amount of
column metadata would be rather large). taplint was not taking
account of this and so generating large numbers of E-TMC-CM21,
and to a lesser extent E-TMC-FM21 reports where services
were taking advantage of this freedom.  Pat Dowler pointed out
out this issue for CADC TAP services.

Fix it so that taplint (a) requests detail=max for VOSI1.1 services,
meaning the service gets a (non-binding) request to supply
per-column metadata from the /tables endpoint, and (b) avoids
issuing the metadata mismatch errors if such detailed metadata
is not supplied.