ttools: rework StatsFilter for multithreading
Various changes to the way that statistics, and especially quantiles,
are calculated for the stilts StatsFilter.  Some of these classes are
used by the topcat StatsWindow as well.

There is now a -[no]parallel flag available for the stats filter
(parallel calculation by default), and the topcat Statistics Window
works in parallel where multiple cores are available.

QuantCalc is replaced with Quantiler, which is in some ways more
complicated (has to allocate memory in a way which admits of
instance merging to support splittable multithreaded processing)
but which also doesn't try to be quite so clever (different approaches
to quantile calculation for different data types).
The current implementation may be slightly slower and slightly
less efficient on memory, but there's probably not much in it
(except for small-valued integers, which the old one would have been
more memory efficient on).

Quantiler is pluggable, which means that alternative implementations
be introduced in future.  At this commit, the only implementation is
SortQuantiler, which does exact calculations.

The quantile calculation semantics are a bit different now:
the results are always floating point values regardless of the
input data type, and the results are interpolated between values
where no exact result is possible.  I think this is more respectable.