ttools: new Quantiler implementation based on GKArray
A new Quantiler implementation is introduced based on the method of
Greenwald and Kanna
(http://infolab.stanford.edu/~datar/courses/cs361a/papers/quantiles.pdf)
via the implementation in https://github.com/DataDog/sketches-java.
The GKArray class was originally com.datadoghq.sketch.gk.GKArray.

This method uses fixed memory and a precision guarantee that I
don't understand (I don't think it's that complicated, I just
can't see what the paper is claiming).  It is amenable to
parallelisation (parallelisation option is provided here).
It is slower in practice than the QuantCalc class I was using before,
but it doesn't run out of memory.

At this commit, the new implementation is not used for quantile
calculations by the STILTS or TOPCAT applications.