table: fix exceptions in table streaming
StreamTableSink and OnceRowPipe/OnceRowPipe2 support
streaming an input table directly to a sink.  They work OK, but
interruption was not working well: what you want is for an
exception thrown in the scanTable abstract method to stop processing.
It was doing that, but leaving some locked read/write threads hanging.
This is complicated, but it's at least partly to do with the fact
that the blocking queue is typically full up when an error is signalled,
so the writes were blocked waiting for reads that never happen.

I've made some changes so this at least behaves better.
I'm not certain I've fixed it for all cases.

I've also switched StreamTableSink to use OnceRowPipe2 rather than
OnceRowPipe.  They are supposed to behave the same, but OnceRowPipe2
may be more reliable (uses java.util.concurrent classes rather than
home-grown wait/notify calls).