ttools: write RandomJELEvaluator
This class is used for evaluating JEL expressions, but it's much
better for parallel use than RandomJELRowReader because it avoids
synchronized contention for thread-bound resources.
Where concurrent use is required, a ThreadLocal is used instead
of synchronization.  This leads to much faster execution in multi-threaded
contexts.

RandomJELRowReader should probably be phased out everywhere in favour
of this class.