topcat: add TopcatJELEvaluator
This class can evaluate JEL expressions and is safe for use
from multiple threads concurrently.  It achieves this by using
ThreadLocals not synchronization, so should be much better for
contended parallel use.

It should therefore be used instead of TopcatJELRowReader
in cases where it's likely to be be called (a lot) from multiple
threads concurrently.