topcat: history note serious error fixed by JEL upgrade
While doing some tests on saved sessions, I noticed that prior to
the recent JEL upgrade to 2.1.2, not only did attempting to reload
boolean synthetic columns sometimes throw an uncaught exception,
but sometimes it silently loaded the column with values inverted
(true for false and vice versa).  This is fixed now, but it
deserves mention in the change log in SUN/253.