topcat: add experimental parallel read plot cache option
Add a toggle to the StackPlotWindow that can configure the
DataStoreFactory to read data in parallel.
Doing it in parallel (generally) makes the plot faster to set up,
but has some problems:

   - it's a bit slower to read from the cache (interact with the plot)
   - cached values may be out of sequence (doesn't matter for most plots)
   - interrupting the read doesn't work properly

The interruption issue is the most serious.  It's probably possible
to fix it, but needs some thought.

Because of these problems, sequential cache preparation is still
the default, and the parallel option hidden in the Plot menu
(with a dumb icon).  But it's there if people want to play around
with very large datasets and think it's worth while.