table: StarTable implements java.io.Closeable
The StarTable interface now includes the close() method.
In most cases this isn't required or useful, but in some cases
there are non-memory resources associated with a table;
case in point is a fits.BintableStarTable which keeps file
descriptors as well as possibly mapped buffers on hand
as long as the table is active.

In most cases this new method can be ignored by implementations
(AbstractStarTable provides a no-op implementation) and by clients,
but for cases where a loop is performed over many tables using it
can be required to avoid e.g. running out of file descriptors.