ttools: don't leak file descriptors in tcat
CatMapper, hence the tcat command, now closes lazily acquired tables
after use, so that the file descriptor limit should not be hit
when attempting to concatenate very large numbers of input tables,
at least when lazy=true.

This uses the new Closeable interface on StarTable.

Other stilts commands could and should be updated in a similar way,
but at least for standalone stilts usage (rather than e.g. within
jystilts) tcat is the only one where this limit is actually likely
to cause trouble (and has done in the past).