topcat: fix MacOS id issue in startup scripts
The topcat and stilts startup scripts need to determine whether
they are running on a Mac to find out where the jar files are
likely to be.  I was using the output of /usr/bin/sw_vers for this,
but it gives a different result on a "Mac Mini" (M1 architecture??)
than the Macs I'd tested it on.  Change the grepping to catch both
types of output.

Patrick Broos spotted and helped to diagnose this issue.