ttools: tweak MapEnvironment input table resolution
Some code in MapEnvironment, under some circumstances, tries to
interpret a table specification as a resource URL relative to
the MapEnvironment class.  If it tries to do that but the resource
does not exist, a NullPointerException results.

Honestly I don't know what use case this is supposed to serve;
quite possibly it should just be removed, but in case it's doing
something important I've left it there, but added a check that
the resource actually exists, and if not fall back to using
the input table string in a more obvious way.