xdoc: rewrite IMG element width/height detection
The code in xhtml.xslt which determines the width and height attributes
for IMG elements in output HTML is rewritten so that it works.
It was previously based on manipulating javax.swing.ImageIcon objects
within the XSLT, and it used to work, but it seems to have broken
for some reason in recent times, probably during the upgrade to
Java 8.  I don't understand the reason for the issue (XAlan XSLT
processing has no access to Swing classes maybe?) but anyway I've
rewritten it with more of the functionality in the java XdocUtils class.
Now it works, in Oracle Java 8 at least.