ttools: fix servlet to work with schemes
The stilts server was ignoring scheme-type table locations if the
tablefactory parameter was set to a non-default value.
Fix it so that it passes schemes.

The code here is a bit of a mess, since it's hard to specify a
table factory with a string.  It may need a more comprehensive
cleanup in the future.