ttools: cope with positionless data plots
Relax some assumptions about how plotters are constructed.
Allow plotters with data-like points, but no actual positions
and no DataGeom.  This is going to support array-valued positions.

I'm not certain this is the best way to do it; possibly there should
be a DataGeom corresponding to X and Y positional vectors, but
I don't think so because the notion of a single position on the
plot surface corresponding to a given row is not very helpful
for a vector of positions.  But there may be unhelpful fallout
from this change, so look out.