ttools: add ObsTapStage UCD/VOUnits unit test
Add a unit test to check that the units and UCDs declared for ObsTAP
model columns are legal according to the VOUnits and UCD standards
respectively.

Note this is really a test of the standard itself rather than the code,
assuming that the code has faithfully copied the units and UCDs from
the documents in question.  Illegal UCDs and Units making their way
into IVOA standard drafts and even RECs is not unknown.