ttools: update taplint VODataService.xsd to v1.2
Update the VODataService XSD to (a PR version of) version 1.2 of
the VODataService standard.  This is required to accommodate
some V1.2 features such as use of the nrows attribute on tables.

This means that all VODataService documents, including pre-V1.2
instances, will be validated against the V1.2 schema (they both have
the same target namespace, "http://www.ivoa.net/xml/VODataService/v1.1";
see the IVOA Schema Versioning note).
It's very hard to do otherwise, since instance documents don't flag
the schema version that they are implementing (and it would be quite
hard even if they did).  That shouldn't matter since it's a minor
update, so any document that validates against v1.1 ought to
validate against v1.2 as well.  Markus assures me that he's confident
this will be the case in practice as well as in theory.

The VODataService-v1.2.xsd document is from volute:
https://volute.g-vo.org/svn/trunk/projects/registry/VODataService/VODataService-v1.2.xsd
at revision 5931.