parquet: add dependency jars
Add jar files required at build and run time by the starjava PARQUET
package.  The parquet files come from parquet-mr, and the others
are a subset of the dependencies declared by parquet-mr/parquet-cli.

To acquire the full set of jar files:

  - clone https://github.com/apache/parquet-mr,
  - checkout tag apache-parquet-1.11.1
  - cd parquet-cli
  - mvn clean install -DskipTests   // probably optional
  - mvn javadoc:javadoc             // optional
  - mvn dependency:copy-dependencies

and the relevant jar files are then in parquet-cli/target and
parquet-cli/target/dependency.

Not all the declared parquet-cli dependencies are needed by the STIL
parquet I/O handlers, and we're trying to cut down the bloat as
much as possible, so some trial and error was used to determine
the minimal set of library jars that were required, which are those
in this commit.

But there's still 15Mb here, which at time of writing is the same size
as the rest of STILTS and all of its depencencies.  This is annoying.

I investigated what could be done to reduce this to only those
classes actually used, by attempting shrinkage of the libraries
based on static dependencies.  The industry standard for this
sort of thing seems to be ProGuard, but that's huge and slow and
complicated and does other things, so I used JarShrink instead:
https://github.com/Deconimus/JarShrink.

I had a go at the largest jar files to see what could be done,
shrinking and adding back classes reported in NoClassDefFoundErrors
until the basic unit tests (of the code added in later commits)
stopped failing:

   % ls -s1rt | sort -n | tail -6 | tac
   3400 hadoop-common-2.7.3.jar
   2684 guava-27.0.1-jre.jar
   1976 snappy-java-1.1.7.3.jar
   1748 parquet-jackson-1.11.1.jar
   1584 parquet-column-1.11.1.jar
    832 parquet-encoding-1.11.1.jar

Taking those in turn:

 - Hadoop-common: reduce 3400k -> 1232k
    % java -jar jarshrink.jar \
           hadoop-common-2.7.3.jar \
        -o hadoop-common-2.7.3-strip.jar \
        -k org.apache.hadoop.conf.Configuration \
        -k org.apache.hadoop.conf.Path \
        -k org.apache.hadoop.security.JniBasedUnixGroupsMappingWithFallback \
        -k org.apache.hadoop.io.compress.CompressionInputStream \
        -k org.apache.hadoop.io.compress.CompressorStream \
        -k org.apache.hadoop.io.compress.DecompressorStream \
        -n

 - Guava: reduce 2684k -> 1420k
    % java -jar jarshrink.jar \
           guava-27.0.1-jre.jar \
        -o guava-27.0.1-jre-strip.jar \
        -k com.google.common.base.Preconditions \
        -k com.google.common.collect.Interners \
        -k com.google.common.base.Ticker \
        -k com.google.common.cache.CacheLoader \
        -k com.google.common.cache.CacheBuilder \
        -n

 - Snappy: it's nearly all arch-specific native libraries, so can't
   strip anything without impacting behaviour on some architectures
   (though admittedly, most are pretty special-interest).

 - Parquet-jackson: I could only reduce 1748k -> 1684k, not worth it.

 - Parquet-column: don't want to risk it, since it may have useful stuff
   not caught by tests.

So I could knock out 3-4Mb out of 15 if I wanted, maybe a bit more if
I do violence to snappy, though with the risk that some of those
classes that get stripped are accessed by reflection or whatever
and will come back to bite me later.

I don't think it's really worth it.  So for now stick with this 15Mb
of support libraries.