parquet: modify parquet-hadoop-1.11.1.jar jar file
The dependency jar file parquet-hadoop-1.11.1.jar contains some
of the same classes as parquet-column-1.11.1.jar
(entries in shaded/parquet/it/unimi/dsi/fastutil).
I think that's probably a parquet-mr build/packaging bug?
Or maybe not, but in any case it causes trouble when trying to combine
jar files with SuperJar, so replace the parquet-hadoop jar file
with one excluding the offending duplicate entries.