ttools: improve CacheFilter
The STILTS CacheFilter now makes sure that there is no reference
in the output table to the input table.  Actually I don't know what
it was before, but it seems there was one.  That means it can now
be used in (jy)stilts scripts like a write/read cycle to reset the
nesting count of wrapped tables; if this isn't done following very
many table filter/wrap operations, StackOverflowExceptions can result.