parquet: add unused FileInputFile implementation
The class FileInputFile implements parquet InputFile without using
Hadoop common classes.  As such it sounds promising, but we still
can't get rid of the (huge) Hadoop common classes, because small
parts of hadoop-common (Path) are used for ParquetWriter.

So this class is not used, but it might possibly come in useful
in the future.